+++ CHANGELOG.md
@@ -8,6 +8,7 @@ We follow [Semantic Versions](https://semver.org/).
 - Implement the post submission to the `RedditCutifier`
 - Add the post submission via the `RedditCutifier` to the `App`
 - Add the `ConfigStorage` class - storing the refresh token
+- Implement the storing of the refresh token to the `Config`
 
 
 ## Version 0.1.4